requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.4
django-cors-headers==2.2.0
django-filter==1.1.0
django-netfields==0.8
djangorestframework==3.8.2
idna==2.6
Markdown==2.6.11
netaddr==0.7.19
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
requests==2.18.4
urllib3==1.22
